feat: introduce @raam/core
