+++ packages/core/src/elements.ts
@@ -1,0 +1,31 @@
+import { String } from "./types/utils";
+
+export type DetermineChildProp =
+  | "ul"
+  | "ol"
+  | "span"
+  | "p"
+  | "h1"
+  | "h2"
+  | "h3"
+  | "h4"
+  | "h5"
+  | "h6"
+  | String
+  | ({} & unknown);
+
+export const determineChild = (parent: DetermineChildProp) =>
+  typeof parent === "string"
+    ? {
+        ul: "li",
+        ol: "li",
+        span: "span",
+        p: "span",
+        h1: "span",
+        h2: "span",
+        h3: "span",
+        h4: "span",
+        h5: "span",
+        h6: "span",
+      }[parent] || "div"
+    : parent;
