+++ packages/core/src/types/csstype.ts
@@ -1,83 +1,4 @@
-export enum FLEX_GAP_CSS_VARS {
-  FLEX_GAP_OFFSET = "--raam-fg-offset",
-  FLEX_GAP = "--raam-fg",
-  FLEX_GAP_TOP = "--raam-fg-t",
-  FLEX_GAP_LEFT = "--raam-fg-l",
-}
-
-/**
- * CSS Style Object
- */
-export type RaamCSS =
-  | CSSProperties
-  | {
-      [
-        /**
-         * **CSS Selector**
-         *
-         * @see https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_Selectors
-         */
-        selector: string
-      ]: RaamCSS;
-    };
-
-type RaamStyleProp<K extends keyof CSSProperties, T = void> =
-  | CSSProperties[K]
-  | T
-  | {
-      [
-        /**
-         * A **`@media`** CSS at-rule can be used to apply the property based on
-         * the result of one or more media queries. With it, you specify a media
-         * query and a style value to apply to the document if and only if the
-         * media query matches the device on which the content is being used.
-         *
-         * **Syntax**: `{ <media-query>: <value> }`
-         *
-         * **Example**: `{ "@media (min-width: 520px)": "1rem" }`
-         *
-         * @see https://developer.mozilla.org/en-US/docs/Web/CSS/@media
-         */
-        mediaQuery: string
-      ]: CSSProperties[K] | T;
-    };
-
-type RaamStylePropsAll<T = void> = {
-  [K in keyof CSSProperties]: RaamStyleProp<K, T> | RaamStyleProp<K, T>[];
-};
-
-/**
- * Responsive Style Props
- */
-export interface RaamStyleProps extends RaamStylePropsAll {}
-
-/**
- * Default RaamTheme
- */
-type ResponsiveThemeValue = StringNumberArr | { [key: string]: StringNumber };
-
-export interface RaamTheme {
-  breakpoints?: ResponsiveThemeValue;
-  space?: ResponsiveThemeValue;
-}
-
-const t: RaamTheme = {
-  breakpoints: [0, 2, 4, 5],
-  // space: [0, "2"],
-  space: { large: 3 },
-};
-
-/**
- * Allow `string` fallbacks without breaking autocomplete
- */
-type String = string & {};
-/**
- * Allow `number` fallbacks without breaking autocomplete
- */
-type Number = number & {};
-
-type StringNumber = string | number;
-type StringNumberArr = StringNumber | StringNumber[];
+import { String, Number } from "./utils";
 
 // csstype v3.0.2
 //
@@ -144,7 +65,7 @@ type SelfPosition =
   | "self-start"
   | "start";
 
-interface CSSProperties {
+export interface CSSProperties {
   /**
    * The CSS **`align-items`** property sets the `align-self` value on all direct children as a group. The align-self property sets the alignment of an item within its containing block. In Flexbox it controls the alignment of items on the Cross Axis, in Grid Layout it controls the alignment of items on the Block Axis within their grid area.
    *
