+++ packages/raam/src/index.ts
@@ -1,2 +1,1 @@
 export * from "./flex";
-export * from "./types";
