+++ packages/tailwind/package.json
@@ -35,6 +35,7 @@
     "tailwindcss": ">= 1.9.2"
   },
   "dependencies": {
+    "@raam/core": "^0.2.3",
     "raam": "^0.2.3"
   },
   "devDependencies": {
