+++ packages/theme-ui/src/flex.tsx
@@ -1,35 +1,20 @@
 import * as React from "react";
 import { Box, useThemeUI } from "theme-ui";
+import { DetermineChildProp, determineChild } from "@raam/core";
 import { useFlex, UseFlexProps, UseFlexChildProps } from "raam";
 
 export interface FlexProps
   extends UseFlexProps,
     Omit<UseFlexChildProps, "index"> {
   children?: React.ReactNode;
-  as?: unknown;
+  as?: DetermineChildProp;
   sx?: any;
 }
-
-const determineChild = (parent: FlexProps["as"]) =>
-  typeof parent === "string"
-    ? {
-        ul: "li",
-        ol: "li",
-        span: "span",
-        p: "span",
-        h1: "span",
-        h2: "span",
-        h3: "span",
-        h4: "span",
-        h5: "span",
-        h6: "span",
-      }[parent] || "div"
-    : parent;
 
 export const Flex = React.forwardRef<HTMLDivElement, FlexProps>(
   (
     {
-      as = "div",
+      as,
       children,
       // flex parent
       alignItems,
